requirements.txt
langchain==0.2.0
langchain-community==0.2.0
langchain-google-genai==1.0.8
faiss-cpu==1.8.0
PyPDF2==3.0.1
pdfplumber==0.11.2
python-telegram-bot==21.4
python-dotenv==1.0.1
google-generativeai==0.7.2
numpy==1.26.4